+++ src/design/plone/policy/tests/test_limit_submit_form.py
@@ -1,6 +1,5 @@
-from collective.volto.formsupport.testing import (  # noqa: E501,
-    VOLTO_FORMSUPPORT_API_FUNCTIONAL_TESTING,
-)
+from design.plone.policy.testing import DESIGN_PLONE_POLICY_API_FUNCTIONAL_TESTING
+from freezegun import freeze_time
 from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import SITE_OWNER_NAME
@@ -10,12 +9,14 @@ from plone.restapi.testing import RelativeSession
 from Products.MailHost.interfaces import IMailHost
 from zope.component import getUtility
 
+import csv
+import io
 import transaction
 import unittest
 
 
 class TestLimitMailStore(unittest.TestCase):
-    layer = VOLTO_FORMSUPPORT_API_FUNCTIONAL_TESTING
+    layer = DESIGN_PLONE_POLICY_API_FUNCTIONAL_TESTING
 
     def setUp(self):
         self.app = self.layer["app"]
@@ -28,8 +29,8 @@ class TestLimitMailStore(unittest.TestCase):
         self.api_session = RelativeSession(self.portal_url)
         self.api_session.headers.update({"Accept": "application/json"})
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
-        self.anon_api_session = RelativeSession(self.portal_url)
-        self.anon_api_session.headers.update({"Accept": "application/json"})
+        # self.anon_api_session = RelativeSession(self.portal_url)
+        # self.anon_api_session.headers.update({"Accept": "application/json"})
 
         self.document = api.content.create(
             type="Document",
@@ -41,19 +42,21 @@ class TestLimitMailStore(unittest.TestCase):
             "text-id": {"@type": "text"},
             "form-id": {"@type": "form"},
         }
-        self.document_url = self.document.absolute_url()
+        self.document_url = (
+            f"{self.portal.absolute_url()}/++api++/{self.document.getId()}"
+        )
         transaction.commit()
 
     def tearDown(self):
         self.api_session.close()
-        self.anon_api_session.close()
+        # self.anon_api_session.close()
 
         # set default block
-        self.document.blocks = {
-            "text-id": {"@type": "text"},
-            "form-id": {"@type": "form"},
-        }
-        transaction.commit()
+        # self.document.blocks = {
+        #     "text-id": {"@type": "text"},
+        #     "form-id": {"@type": "form"},
+        # }
+        # transaction.commit()
 
     def submit_form(self, data):
         url = f"{self.document_url}/@submit-form"
@@ -61,9 +64,9 @@ class TestLimitMailStore(unittest.TestCase):
             url,
             json=data,
         )
-        transaction.commit()
         return response
 
+    @freeze_time("2025-12-29T23:09:33")
     def test_limit_submit(self):
         self.document.blocks = {
             "form-id": {
@@ -98,7 +101,6 @@ class TestLimitMailStore(unittest.TestCase):
                 "block_id": "form-id",
             },
         )
-        transaction.commit()
         self.assertEqual(response.status_code, 200)
 
         response = self.submit_form(
@@ -113,9 +115,34 @@ class TestLimitMailStore(unittest.TestCase):
                 "block_id": "form-id",
             },
         )
-        transaction.commit()
         self.assertEqual(response.status_code, 200)
+
         self.assertTrue(response.json()["waiting_list"])
+
+        # export csv
+        response = self.api_session.get(f"{self.document_url}/@form-data-export")
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(
+            response.headers["Content-Type"],
+            "text/comma-separated-values; charset=utf-8",
+        )
+        self.assertEqual(
+            [r for r in csv.DictReader(io.StringIO(response.text))],
+            [
+                {
+                    "Message": "just want to say hi",
+                    "Name": "John",
+                    "Lista d'attesa": "No",
+                    "date": "2025-12-29T23:09:33",
+                },
+                {
+                    "Message": "just want to say hi",
+                    "Name": "John",
+                    "Lista d'attesa": "Si",
+                    "date": "2025-12-29T23:09:33",
+                },
+            ],
+        )
 
     def test_unique_field(self):
         self.document.blocks = {
@@ -151,14 +178,14 @@ class TestLimitMailStore(unittest.TestCase):
         }
 
         response = self.submit_form(data=data)
-        transaction.commit()
-
         self.assertEqual(response.status_code, 200)
 
         response = self.submit_form(data=data)
-        transaction.commit()
+        self.assertEqual(response.status_code, 400)
 
-        self.assertEqual(response.status_code, 500)
         # test message is not fair because it's a translation, in another package
         message = response.json()["message"]
-        self.assertTrue("Value not unique" in message or "non sono univoci" in message)
+        self.assertEqual(
+            message,
+            "[{'message': 'Unable to save data. The value of field \"Name\" is already stored in previous submissions.', 'field_id': 'name', 'label': 'Name', 'error': 'UniqueValueError'}]",
+        )