+++ src/design/plone/policy/tests/test_setup.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """Setup tests for this package."""
+
 from design.plone.policy.testing import DESIGN_PLONE_POLICY_INTEGRATION_TESTING
 from design.plone.policy.testing import (
     DESIGN_PLONE_POLICY_LIMIT_ROOT_ADDABLES_INTEGRATION_TESTING,
@@ -14,7 +15,6 @@ from Products.CMFPlone.interfaces.controlpanel import INavigationSchema
 from zope.component import getUtility
 
 import unittest
-
 
 try:
     from Products.CMFPlone.utils import get_installer